+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Viewer: allow setting default outside video player
+- Map: fit to most recent items if all items cannot fit on screen
 
 ## <a id="v1.7.7"></a>[v1.7.7] - 2022-11-27